Extract and fix viewing key tests
